feat: db 로직 클래스 분리
